requirements.txt
PyOpenGL
PySDL2
PIL
numpy
epydoc
